IMPROVE: Check cluster health
Before attempting to execute the next scripts, we need to check if the K8S cluster is healthy.

Motivation: My selected zone had issues with cluster creation for about 10 minutes and my scripts were unable to create the resources. Some commands ran out of order or it just failed at different parts.

This commit executes to check if the cluster is really ready.